Changed side effect id to bigint in relationship table and changed schema.rb

+++ db/migrate/20190611100029_create_side_effects_has_medication_prescriptions.rb
@@ -3,7 +3,7 @@
 class CreateSideEffectsHasMedicationPrescriptions < ActiveRecord::Migration[5.2]
   def change
     create_table :side_effects_has_medication_prescriptions, primary_key: :side_effects_has_medication_prescription_id do |t|
-      t.integer :side_effect_id, null: false
+      t.bigint :side_effect_id, null: false
       t.integer :medication_prescription_id, null: false
 
       t.timestamps
